moved scape x mobile to extensions folder
